Adapt imports after gammapy.utils.gauss move
