+++ gammapy/cube/psf_kernel.py
@@ -5,8 +5,8 @@ from astropy.coordinates import Angle
 from astropy.coordinates.angle_utilities import angular_separation
 import astropy.units as u
 from astropy.convolution import convolve_fft
+from ..utils.gauss import Gauss2DPDF
 from ..maps import Map, WcsGeom
-from ..image.models.gauss import Gauss2DPDF
 from ..irf import TablePSF
 
 __all__ = [
