+++ gammapy/irf/psf_core.py
@@ -19,7 +19,7 @@ import json
 import numpy as np
 from astropy.convolution import Gaussian2DKernel
 from astropy.stats import gaussian_fwhm_to_sigma, gaussian_sigma_to_fwhm
-from ..image.models.gauss import Gauss2DPDF, MultiGauss2D
+from ..utils.gauss import Gauss2DPDF, MultiGauss2D
 
 __all__ = [
     'GaussPSF',
